Refactor MediaCard components to use SortableBookshelfOverlayContext
